Request feature: Proximity Filter
the map already has various filters, in sections which are collapsed and expanded using jquery.  Proximity search would be a new tool added to the "advanced" filters.  It would allow the user to view only services within a certain radius of either their current location, or a location which they specify with a tap/click on the map.