Feature Request: Region Filter
given a geoJSON file of polygons which correspond to the regions of Jordan, provide a filter in Advanced Search which would let the user either select one region from a list, or tap/click the region on the map.  Then filter the markers to ones which are located inside the selected region.